backend/requirements.txt
fastapi
uvicorn
uvicorn[standard]
pydantic-settings
langchain
langchain-cohere
langchain-text-splitters
cohere
faiss-cpu
networkx
langfuse
unstructured
unstructured[pdf]
numpy
PyMuPDF
python-docx
openpyxl
sqlalchemy
python-multipart
sse-starlette
